# perf: avoid gumbo_debug function calls when GUMBO_DEBUG isn't def'd