**What problem is this PR intended to solve?**

Part of #2722, optimizing performance of gumbo.

Benchmarks from running show 7-11% improvements.

-  "test/files/GH_1042.html", # 650b
-  "test/files/tlm.html", # 70kb
-  "big_shopping.html", # 1.9mb

```
Comparison:
nodbg html5 parse 656:    16646.6 i/s
BASE_ html5 parse 656:    15545.9 i/s - same-ish: difference falls within error
                   with 99.0% confidence

Comparison:
nodbg html5 parse 70095:      271.1 i/s
BASE_ html5 parse 70095:      253.1 i/s - same-ish: difference falls within error
                   with 99.0% confidence

Comparison:
nodbg html5 parse 1929522:       14.9 i/s
BASE_ html5 parse 1929522:       13.4 i/s - 1.11x  (± 0.07) slower
                   with 99.0% confidence
```